LayoutTests/platform/mac-wk2/http/tests/websocket/tests/hybi/send-object-tostring-check-expected.txt

WebSocket: Object's toString method should be called only once.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testObj.callCounter is 1
FAIL closeEvent.wasClean should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
